Implement slots game
This commit adds a new navigational item to the nav bar as well as a new view for
replacing the casino-content turbo frame. It implements a slot machine style game that randomly choses three items from a list and displays them to the user. Specific rules are set up in the controller to handle win conditions on each playy of the game and modifies the current session funds appropriately.